+++ api/search.js
@@ -7,11 +7,21 @@ module.exports = async function handler(req, res) {
     const page = parseInt((req.query.page || '0').toString(), 10) || 0;
     const perPage = Math.min(100, parseInt((req.query.per_page || '20').toString(), 10) || 20);
 
-    const terms = q.split(/\s+/).map(s => s.trim().toLowerCase()).filter(w => w.length > 2);
-    if (!terms.length) return res.status(200).json({ success: true, data: { profiles: [], total_count: 0, page, per_page: perPage }, error: null });
+    const terms = q
+      .split(/\s+/)
+      .map((s) => s.trim().toLowerCase())
+      .filter((w) => w.length > 2);
+    if (!terms.length)
+      return res
+        .status(200)
+        .json({
+          success: true,
+          data: { profiles: [], total_count: 0, page, per_page: perPage },
+          error: null
+        });
 
     const where = terms.map(() => '(si.term LIKE ? OR p.search_vector LIKE ?)').join(' OR ');
-    const whereArgs = terms.flatMap(t => [`%${t}%`, `%${t}%`]);
+    const whereArgs = terms.flatMap((t) => [`%${t}%`, `%${t}%`]);
 
     const client = getClient();
     const countSql = `SELECT COUNT(DISTINCT p.pubkey) AS c FROM profiles p JOIN search_index si ON p.pubkey = si.pubkey WHERE ${where}`;
@@ -22,7 +32,7 @@ module.exports = async function handler(req, res) {
     const listArgs = [...whereArgs, perPage, page * perPage];
     const rows = await client.execute({ sql: listSql, args: listArgs });
 
-    const profiles = rows.rows.map(r => ({
+    const profiles = rows.rows.map((r) => ({
       pubkey: r.pubkey,
       name: r.name || null,
       display_name: r.display_name || null,
@@ -38,7 +48,13 @@ module.exports = async function handler(req, res) {
       search_terms: []
     }));
 
-    res.status(200).json({ success: true, data: { profiles, total_count: total, page, per_page: perPage }, error: null });
+    res
+      .status(200)
+      .json({
+        success: true,
+        data: { profiles, total_count: total, page, per_page: perPage },
+        error: null
+      });
   } catch (e) {
     res.status(500).json({ success: false, data: null, error: e?.message || 'error' });
   }
