+++ db/local-migration-manager.js
@@ -11,13 +11,14 @@ class LocalMigrationManager {
    */
   async runMigrations() {
     const client = this.dbManager.getClient();
-    
+
     return new Promise((resolve, reject) => {
       try {
         console.log('Running database migrations...');
-        
+
         // Create profiles table
-        client.run(`
+        client.run(
+          `
           CREATE TABLE IF NOT EXISTS profiles (
             pubkey TEXT PRIMARY KEY,
             npub TEXT NOT NULL,
@@ -33,15 +34,17 @@ class LocalMigrationManager {
             indexed_at INTEGER NOT NULL,
             search_vector TEXT
           )
-        `, (err) => {
-          if (err) {
-            console.error('Failed to create profiles table:', err);
-            reject(err);
-            return;
-          }
+        `,
+          (err) => {
+            if (err) {
+              console.error('Failed to create profiles table:', err);
+              reject(err);
+              return;
+            }
 
-          // Create relationships table
-          client.run(`
+            // Create relationships table
+            client.run(
+              `
             CREATE TABLE IF NOT EXISTS relationships (
               follower_pubkey TEXT NOT NULL,
               following_pubkey TEXT NOT NULL,
@@ -53,52 +56,61 @@ class LocalMigrationManager {
               indexed_at INTEGER NOT NULL,
               PRIMARY KEY (follower_pubkey, following_pubkey)
             )
-          `, (err) => {
-            if (err) {
-              console.error('Failed to create relationships table:', err);
-              reject(err);
-              return;
-            }
+          `,
+              (err) => {
+                if (err) {
+                  console.error('Failed to create relationships table:', err);
+                  reject(err);
+                  return;
+                }
 
-            // Create search_index table
-            client.run(`
+                // Create search_index table
+                client.run(
+                  `
               CREATE TABLE IF NOT EXISTS search_index (
                 term TEXT NOT NULL,
                 pubkey TEXT NOT NULL,
                 field_type TEXT NOT NULL,
                 PRIMARY KEY (term, pubkey, field_type)
               )
-            `, (err) => {
-              if (err) {
-                console.error('Failed to create search_index table:', err);
-                reject(err);
-                return;
-              }
+            `,
+                  (err) => {
+                    if (err) {
+                      console.error('Failed to create search_index table:', err);
+                      reject(err);
+                      return;
+                    }
 
-              // Create indexer_state table
-              client.run(`
+                    // Create indexer_state table
+                    client.run(
+                      `
                 CREATE TABLE IF NOT EXISTS indexer_state (
                   key TEXT PRIMARY KEY,
                   value TEXT NOT NULL
                 )
-              `, (err) => {
-                if (err) {
-                  console.error('Failed to create indexer_state table:', err);
-                  reject(err);
-                  return;
-                }
+              `,
+                      (err) => {
+                        if (err) {
+                          console.error('Failed to create indexer_state table:', err);
+                          reject(err);
+                          return;
+                        }
 
-                // Create indexes for better performance
-                this.createIndexes(client)
-                  .then(() => {
-                    console.log('Database migrations completed successfully');
-                    resolve(true);
-                  })
-                  .catch(reject);
-              });
-            });
-          });
-        });
+                        // Create indexes for better performance
+                        this.createIndexes(client)
+                          .then(() => {
+                            console.log('Database migrations completed successfully');
+                            resolve(true);
+                          })
+                          .catch(reject);
+                      }
+                    );
+                  }
+                );
+              }
+            );
+          }
+        );
       } catch (error) {
         console.error('Migration failed:', error);
         reject(error);
@@ -125,7 +137,7 @@ class LocalMigrationManager {
       'CREATE INDEX IF NOT EXISTS idx_search_index_field_type ON search_index(field_type)'
     ];
 
-    const createIndexPromises = indexes.map(index => {
+    const createIndexPromises = indexes.map((index) => {
       return new Promise((resolve, reject) => {
         client.run(index, (err) => {
           if (err) {
@@ -147,11 +159,11 @@ class LocalMigrationManager {
    */
   async resetDatabase() {
     const client = this.dbManager.getClient();
-    
+
     return new Promise((resolve, reject) => {
       try {
         console.log('Resetting database...');
-        
+
         client.run('DROP TABLE IF EXISTS search_index', (err) => {
           if (err) {
             reject(err);
@@ -187,43 +199,46 @@ class LocalMigrationManager {
    */
   async getSchemaInfo() {
     const client = this.dbManager.getClient();
-    
+
     return new Promise((resolve, reject) => {
       try {
-        client.all(`
+        client.all(
+          `
           SELECT name FROM sqlite_master 
           WHERE type='table' AND name NOT LIKE 'sqlite_%'
           ORDER BY name
-        `, (err, tables) => {
-          if (err) {
-            reject(err);
-            return;
-          }
+        `,
+          (err, tables) => {
+            if (err) {
+              reject(err);
+              return;
+            }
 
-          const schemaInfo = {};
-          let processedTables = 0;
+            const schemaInfo = {};
+            let processedTables = 0;
 
-          if (tables.length === 0) {
-            resolve(schemaInfo);
-            return;
-          }
+            if (tables.length === 0) {
+              resolve(schemaInfo);
+              return;
+            }
 
-          tables.forEach(table => {
-            const tableName = table.name;
-            client.all(`PRAGMA table_info(${tableName})`, (err, columns) => {
-              if (err) {
-                console.warn(`Failed to get columns for table ${tableName}:`, err.message);
-              } else {
-                schemaInfo[tableName] = columns;
-              }
+            tables.forEach((table) => {
+              const tableName = table.name;
+              client.all(`PRAGMA table_info(${tableName})`, (err, columns) => {
+                if (err) {
+                  console.warn(`Failed to get columns for table ${tableName}:`, err.message);
+                } else {
+                  schemaInfo[tableName] = columns;
+                }
 
-              processedTables++;
-              if (processedTables === tables.length) {
-                resolve(schemaInfo);
-              }
+                processedTables++;
+                if (processedTables === tables.length) {
+                  resolve(schemaInfo);
+                }
+              });
             });
-          });
-        });
+          }
+        );
       } catch (error) {
         console.error('Failed to get schema info:', error);
         reject(error);
