+++ db/local-relationship-manager.js
@@ -13,7 +13,7 @@ class LocalRelationshipManager {
    */
   async upsertRelationship(relationship) {
     const client = this.dbManager.getClient();
-    
+
     return new Promise((resolve, reject) => {
       try {
         const followerNpub = hexToNpub(relationship.follower_pubkey);
@@ -25,7 +25,7 @@ class LocalRelationshipManager {
             relay, petname, created_at, indexed_at
           ) VALUES (?, ?, ?, ?, ?, ?, ?, ?)
         `;
-        
+
         const args = [
           relationship.follower_pubkey,
           relationship.following_pubkey,
@@ -37,7 +37,7 @@ class LocalRelationshipManager {
           relationship.indexed_at
         ];
 
-        client.run(sql, args, function(err) {
+        client.run(sql, args, function (err) {
           if (err) {
             console.error('Failed to upsert relationship:', err);
             reject(err);
@@ -57,11 +57,11 @@ class LocalRelationshipManager {
    */
   async getFollowing(pubkey, limit = 100) {
     const client = this.dbManager.getClient();
-    
+
     return new Promise((resolve, reject) => {
       try {
         const hexPubkey = pubkey.startsWith('npub') ? npubToHex(pubkey) : pubkey;
-        
+
         const sql = `
           SELECT r.following_pubkey, r.following_npub, r.relay, r.petname, r.created_at,
                  p.name, p.display_name, p.picture, p.nip05
@@ -77,18 +77,20 @@ class LocalRelationshipManager {
             console.error('Failed to get following:', err);
             reject(err);
           } else {
-            const following = rows.map(row => ({
+            const following = rows.map((row) => ({
               pubkey: row.following_pubkey,
               npub: row.following_npub,
               relay: row.relay,
               petname: row.petname,
               created_at: row.created_at,
-              profile: row.name ? {
-                name: row.name,
-                display_name: row.display_name,
-                picture: row.picture,
-                nip05: row.nip05
-              } : null
+              profile: row.name
+                ? {
+                    name: row.name,
+                    display_name: row.display_name,
+                    picture: row.picture,
+                    nip05: row.nip05
+                  }
+                : null
             }));
             resolve(following);
           }
@@ -105,11 +107,11 @@ class LocalRelationshipManager {
    */
   async getFollowers(pubkey, limit = 100) {
     const client = this.dbManager.getClient();
-    
+
     return new Promise((resolve, reject) => {
       try {
         const hexPubkey = pubkey.startsWith('npub') ? npubToHex(pubkey) : pubkey;
-        
+
         const sql = `
           SELECT r.follower_pubkey, r.follower_npub, r.relay, r.petname, r.created_at,
                  p.name, p.display_name, p.picture, p.nip05
@@ -125,18 +127,20 @@ class LocalRelationshipManager {
             console.error('Failed to get followers:', err);
             reject(err);
           } else {
-            const followers = rows.map(row => ({
+            const followers = rows.map((row) => ({
               pubkey: row.follower_pubkey,
               npub: row.follower_npub,
               relay: row.relay,
               petname: row.petname,
               created_at: row.created_at,
-              profile: row.name ? {
-                name: row.name,
-                display_name: row.display_name,
-                picture: row.picture,
-                nip05: row.nip05
-              } : null
+              profile: row.name
+                ? {
+                    name: row.name,
+                    display_name: row.display_name,
+                    picture: row.picture,
+                    nip05: row.nip05
+                  }
+                : null
             }));
             resolve(followers);
           }
@@ -153,46 +157,56 @@ class LocalRelationshipManager {
    */
   async getRelationshipStats(pubkey) {
     const client = this.dbManager.getClient();
-    
+
     return new Promise((resolve, reject) => {
       try {
         const hexPubkey = pubkey.startsWith('npub') ? npubToHex(pubkey) : pubkey;
-        
-        // Get following count
-        client.get('SELECT COUNT(*) as count FROM relationships WHERE follower_pubkey = ?', 
-          [hexPubkey], (err, followingRow) => {
-          if (err) {
-            reject(err);
-            return;
-          }
 
-          // Get followers count
-          client.get('SELECT COUNT(*) as count FROM relationships WHERE following_pubkey = ?', 
-            [hexPubkey], (err, followersRow) => {
+        // Get following count
+        client.get(
+          'SELECT COUNT(*) as count FROM relationships WHERE follower_pubkey = ?',
+          [hexPubkey],
+          (err, followingRow) => {
             if (err) {
               reject(err);
               return;
             }
 
-            // Get last contact update
-            client.get(`
+            // Get followers count
+            client.get(
+              'SELECT COUNT(*) as count FROM relationships WHERE following_pubkey = ?',
+              [hexPubkey],
+              (err, followersRow) => {
+                if (err) {
+                  reject(err);
+                  return;
+                }
+
+                // Get last contact update
+                client.get(
+                  `
               SELECT MAX(created_at) as last_update 
               FROM relationships 
               WHERE follower_pubkey = ? OR following_pubkey = ?
-            `, [hexPubkey, hexPubkey], (err, lastUpdateRow) => {
-              if (err) {
-                reject(err);
-                return;
-              }
+            `,
+                  [hexPubkey, hexPubkey],
+                  (err, lastUpdateRow) => {
+                    if (err) {
+                      reject(err);
+                      return;
+                    }
 
-              resolve({
-                following_count: followingRow.count,
-                followers_count: followersRow.count,
-                last_contact_update: lastUpdateRow.last_update
-              });
-            });
-          });
-        });
+                    resolve({
+                      following_count: followingRow.count,
+                      followers_count: followersRow.count,
+                      last_contact_update: lastUpdateRow.last_update
+                    });
+                  }
+                );
+              }
+            );
+          }
+        );
       } catch (error) {
         console.error('Failed to get relationship stats:', error);
         reject(error);
@@ -205,12 +219,16 @@ class LocalRelationshipManager {
    */
   async isFollowing(followerPubkey, followingPubkey) {
     const client = this.dbManager.getClient();
-    
+
     return new Promise((resolve, reject) => {
       try {
-        const followerHex = followerPubkey.startsWith('npub') ? npubToHex(followerPubkey) : followerPubkey;
-        const followingHex = followingPubkey.startsWith('npub') ? npubToHex(followingPubkey) : followingPubkey;
-        
+        const followerHex = followerPubkey.startsWith('npub')
+          ? npubToHex(followerPubkey)
+          : followerPubkey;
+        const followingHex = followingPubkey.startsWith('npub')
+          ? npubToHex(followingPubkey)
+          : followingPubkey;
+
         const sql = `
           SELECT COUNT(*) as count 
           FROM relationships
