+++ db/migration-manager.js
@@ -11,10 +11,10 @@ class MigrationManager {
    */
   async runMigrations() {
     const client = this.dbManager.getClient();
-    
+
     try {
       console.log('Running database migrations...');
-      
+
       // Create profiles table
       await client.execute(`
         CREATE TABLE IF NOT EXISTS profiles (
@@ -69,7 +69,7 @@ class MigrationManager {
 
       // Create indexes for better performance
       await this.createIndexes(client);
-      
+
       console.log('Database migrations completed successfully');
       return true;
     } catch (error) {
@@ -111,14 +111,14 @@ class MigrationManager {
    */
   async resetDatabase() {
     const client = this.dbManager.getClient();
-    
+
     try {
       console.log('Resetting database...');
-      
+
       await client.execute('DROP TABLE IF EXISTS search_index');
       await client.execute('DROP TABLE IF EXISTS relationships');
       await client.execute('DROP TABLE IF EXISTS profiles');
-      
+
       console.log('Database reset completed');
     } catch (error) {
       console.error('Database reset failed:', error);
@@ -131,22 +131,22 @@ class MigrationManager {
    */
   async getSchemaInfo() {
     const client = this.dbManager.getClient();
-    
+
     try {
       const tables = await client.execute(`
         SELECT name FROM sqlite_master 
         WHERE type='table' AND name NOT LIKE 'sqlite_%'
         ORDER BY name
       `);
-      
+
       const schemaInfo = {};
-      
+
       for (const table of tables.rows) {
         const tableName = table.name;
         const columns = await client.execute(`PRAGMA table_info(${tableName})`);
         schemaInfo[tableName] = columns.rows;
       }
-      
+
       return schemaInfo;
     } catch (error) {
       console.error('Failed to get schema info:', error);
