+++ db/relationship-manager.js
@@ -13,7 +13,7 @@ class RelationshipManager {
    */
   async upsertRelationship(relationship) {
     const client = this.dbManager.getClient();
-    
+
     try {
       const followerNpub = hexToNpub(relationship.follower_pubkey);
       const followingNpub = hexToNpub(relationship.following_pubkey);
@@ -50,10 +50,10 @@ class RelationshipManager {
    */
   async getFollowing(pubkey, limit = 100) {
     const client = this.dbManager.getClient();
-    
+
     try {
       const hexPubkey = pubkey.startsWith('npub') ? npubToHex(pubkey) : pubkey;
-      
+
       const result = await client.execute({
         sql: `
           SELECT follower_pubkey, following_pubkey, follower_npub, following_npub,
@@ -66,7 +66,7 @@ class RelationshipManager {
         args: [hexPubkey, limit]
       });
 
-      return result.rows.map(row => ({
+      return result.rows.map((row) => ({
         follower_pubkey: row.follower_pubkey,
         following_pubkey: row.following_pubkey,
         follower_npub: row.follower_npub,
@@ -87,10 +87,10 @@ class RelationshipManager {
    */
   async getFollowers(pubkey, limit = 100) {
     const client = this.dbManager.getClient();
-    
+
     try {
       const hexPubkey = pubkey.startsWith('npub') ? npubToHex(pubkey) : pubkey;
-      
+
       const result = await client.execute({
         sql: `
           SELECT follower_pubkey, following_pubkey, follower_npub, following_npub,
@@ -103,7 +103,7 @@ class RelationshipManager {
         args: [hexPubkey, limit]
       });
 
-      return result.rows.map(row => ({
+      return result.rows.map((row) => ({
         follower_pubkey: row.follower_pubkey,
         following_pubkey: row.following_pubkey,
         follower_npub: row.follower_npub,
@@ -124,12 +124,18 @@ class RelationshipManager {
    */
   async getRelationshipStats(pubkey) {
     const client = this.dbManager.getClient();
-    
+
     try {
       const hexPubkey = pubkey.startsWith('npub') ? npubToHex(pubkey) : pubkey;
-      
-      const followingRes = await client.execute({ sql: 'SELECT COUNT(*) AS c FROM relationships WHERE follower_pubkey = ?1', args: [hexPubkey] });
-      const followersRes = await client.execute({ sql: 'SELECT COUNT(*) AS c FROM relationships WHERE following_pubkey = ?1', args: [hexPubkey] });
+
+      const followingRes = await client.execute({
+        sql: 'SELECT COUNT(*) AS c FROM relationships WHERE follower_pubkey = ?1',
+        args: [hexPubkey]
+      });
+      const followersRes = await client.execute({
+        sql: 'SELECT COUNT(*) AS c FROM relationships WHERE following_pubkey = ?1',
+        args: [hexPubkey]
+      });
       const lastUpdateRes = await client.execute({
         sql: 'SELECT MAX(indexed_at) AS last_update FROM relationships WHERE follower_pubkey = ?1 OR following_pubkey = ?1',
         args: [hexPubkey]
@@ -139,9 +145,10 @@ class RelationshipManager {
         pubkey: hexPubkey,
         following_count: (followingRes.rows[0] && Number(followingRes.rows[0].c)) || 0,
         followers_count: (followersRes.rows[0] && Number(followersRes.rows[0].c)) || 0,
-        last_contact_update: lastUpdateRes.rows[0] && lastUpdateRes.rows[0].last_update
-          ? new Date(Number(lastUpdateRes.rows[0].last_update) * 1000).toISOString()
-          : null
+        last_contact_update:
+          lastUpdateRes.rows[0] && lastUpdateRes.rows[0].last_update
+            ? new Date(Number(lastUpdateRes.rows[0].last_update) * 1000).toISOString()
+            : null
       };
     } catch (error) {
       console.error('Failed to get relationship stats:', error);
@@ -154,7 +161,7 @@ class RelationshipManager {
    */
   async relationshipExists(followerPubkey, followingPubkey) {
     const client = this.dbManager.getClient();
-    
+
     try {
       const result = await client.execute({
         sql: `
@@ -176,7 +183,7 @@ class RelationshipManager {
    */
   async deleteRelationship(followerPubkey, followingPubkey) {
     const client = this.dbManager.getClient();
-    
+
     try {
       await client.execute({
         sql: `
@@ -198,10 +205,10 @@ class RelationshipManager {
    */
   async getMutualFollowers(pubkey, limit = 50) {
     const client = this.dbManager.getClient();
-    
+
     try {
       const hexPubkey = pubkey.startsWith('npub') ? npubToHex(pubkey) : pubkey;
-      
+
       const result = await client.execute({
         sql: `
           SELECT r1.follower_pubkey, r1.follower_npub
@@ -215,7 +222,7 @@ class RelationshipManager {
         args: [hexPubkey, limit]
       });
 
-      return result.rows.map(row => ({
+      return result.rows.map((row) => ({
         pubkey: row.follower_pubkey,
         npub: row.follower_npub
       }));
@@ -230,16 +237,22 @@ class RelationshipManager {
    */
   async getBulkRelationshipStats(pubkeys) {
     const client = this.dbManager.getClient();
-    
+
     try {
       const stats = [];
-      
+
       for (const pubkey of pubkeys) {
         const hexPubkey = pubkey.startsWith('npub') ? npubToHex(pubkey) : pubkey;
-        
-        const followingRes = await client.execute({ sql: 'SELECT COUNT(*) AS c FROM relationships WHERE follower_pubkey = ?1', args: [hexPubkey] });
-        const followersRes = await client.execute({ sql: 'SELECT COUNT(*) AS c FROM relationships WHERE following_pubkey = ?1', args: [hexPubkey] });
-        
+
+        const followingRes = await client.execute({
+          sql: 'SELECT COUNT(*) AS c FROM relationships WHERE follower_pubkey = ?1',
+          args: [hexPubkey]
+        });
+        const followersRes = await client.execute({
+          sql: 'SELECT COUNT(*) AS c FROM relationships WHERE following_pubkey = ?1',
+          args: [hexPubkey]
+        });
+
         stats.push({
           pubkey: hexPubkey,
           following_count: (followingRes.rows[0] && Number(followingRes.rows[0].c)) || 0,
@@ -259,11 +272,20 @@ class RelationshipManager {
    */
   async getRelationshipStats() {
     const client = this.dbManager.getClient();
-    
+
     try {
-      const totalRes = await client.execute({ sql: 'SELECT COUNT(*) AS c FROM relationships', args: [] });
-      const uniqueFollowersRes = await client.execute({ sql: 'SELECT COUNT(DISTINCT follower_pubkey) AS c FROM relationships', args: [] });
-      const uniqueFollowingRes = await client.execute({ sql: 'SELECT COUNT(DISTINCT following_pubkey) AS c FROM relationships', args: [] });
+      const totalRes = await client.execute({
+        sql: 'SELECT COUNT(*) AS c FROM relationships',
+        args: []
+      });
+      const uniqueFollowersRes = await client.execute({
+        sql: 'SELECT COUNT(DISTINCT follower_pubkey) AS c FROM relationships',
+        args: []
+      });
+      const uniqueFollowingRes = await client.execute({
+        sql: 'SELECT COUNT(DISTINCT following_pubkey) AS c FROM relationships',
+        args: []
+      });
 
       return {
         total_relationships: (totalRes.rows[0] && Number(totalRes.rows[0].c)) || 0,
