+++ package.json
@@ -9,37 +9,21 @@
     "db:stats": "node db/migrate.js stats",
     "db:health": "node db/migrate.js health",
     "db:reset": "node db/migrate.js reset",
-
     "index:backfill:turso": "node scripts/indexer-backfill-turso.js",
     "index:backfill:turso:profiles": "node scripts/indexer-backfill-turso.js --only=profiles",
     "index:backfill:turso:contacts": "node scripts/indexer-backfill-turso.js --only=contacts",
-
     "index:backfill:sqlite": "node scripts/indexer-backfill-sqlite.js",
     "index:backfill:sqlite:profiles": "node scripts/indexer-backfill-sqlite.js --only=profiles",
     "index:backfill:sqlite:contacts": "node scripts/indexer-backfill-sqlite.js --only=contacts",
-
     "index:enhanced:sqlite": "node scripts/indexer-enhanced-sqlite.js",
     "index:enhanced:sqlite:profiles": "node scripts/indexer-enhanced-sqlite.js --only=profiles",
     "index:enhanced:sqlite:contacts": "node scripts/indexer-enhanced-sqlite.js --only=contacts",
-
     "dev:api": "node scripts/dev-api-server.js",
     "sqlite:stats": "node scripts/sqlite-check-stats.js",
     "sqlite:migrate:enhanced": "node scripts/sqlite-migrate-enhanced.js",
     "sqlite:test:enhanced": "node scripts/sqlite-test-enhanced.js",
-
-    "index:local": "node scripts/local-index.js",
-    "index:profiles": "node scripts/local-index.js --only=profiles",
-    "index:contacts": "node scripts/local-index.js --only=contacts",
-    "index:local-sqlite": "node scripts/local-index-sqlite.js",
-    "index:local-profiles": "node scripts/local-index-sqlite.js --only=profiles",
-    "index:local-contacts": "node scripts/local-index-sqlite.js --only=contacts",
-    "index:enhanced": "node scripts/enhanced-indexer.js",
-    "index:enhanced-profiles": "node scripts/enhanced-indexer.js --only=profiles",
-    "index:enhanced-contacts": "node scripts/enhanced-indexer.js --only=contacts",
-    "stats": "node scripts/check-stats.js",
-    "api": "node scripts/fast-query-api.js",
-    "test:enhanced": "node scripts/test-enhanced.js",
-    "migrate:enhanced": "node scripts/migrate-enhanced.js"
+    "lint": "eslint . --max-warnings=0",
+    "format": "prettier --write \"**/*.{js,json,md}\""
   },
   "dependencies": {
     "@libsql/client": "latest",
@@ -49,7 +33,11 @@
     "ws": "^8.17.0"
   },
   "devDependencies": {
-    "dotenv": "^16.0.0"
+    "dotenv": "^16.0.0",
+    "eslint": "^9.34.0",
+    "eslint-config-prettier": "^9.1.2",
+    "eslint-plugin-prettier": "^5.5.4",
+    "prettier": "^3.6.2"
   },
   "engines": {
     "node": ">=18.0.0"
