+++ vercel.json
@@ -13,8 +13,6 @@
   "functions": {
     "api/indexer-cron.js": { "maxDuration": 60, "memory": 1024 }
   },
-  "crons": [
-    { "path": "/api/indexer-cron", "schedule": "0 2 * * *" }
-  ],
+  "crons": [{ "path": "/api/indexer-cron", "schedule": "0 2 * * *" }],
   "regions": ["iad1"]
 }
